Modify creation date spider
In meeting with my reviewer we agree on creation date of a spider, it must be oldest deploy associated with that spider. Additional information of my changes are: modified code for getting oldest creation date of a deploy associated to my spider, delete REACT_APP_API_BASE_URL variable on .env.development file and delete property output in webpack configuration file.