BITMAKER-2199 estela-web: Implement "Spider/Overview" mockup (#112)
* Add List of Cards and List of Rows components

I added the above section of the mockup, the first two Cards and the List of Rows that show general information about the spider, the calls to the API are in progress

* Add SpiderJobs Data and filtering button

Added SpiderJobs additional information such as tags and arguments, checking estela api and status of Jobs to organize different tables for jobs

* Ignore some content in the mockup

In meetings with my reviewer Raymond, we agreed on ignoring some fields that are not required yet in the spider detail mockup (such as related scheduled jobs) or should be implemented in the backend side (such as creation date field in the Spider model)

* Issue on server MIME type error

* Revert "Fix output problem (#110)"

This reverts commit db66b4e008e58ba73ac8532ac6037c42191d4f10.

reverting because I found some issues with compilation of tsx components

* Add Tables logic in overview page

I added the tables logic in the overview page to see queued, running and completed jobs, also in a meeting with my reviewer Raymond we noticed an issue between estelaCLI and estela web interface we will fix it soon.

* Add Spider creation date based on deploy

Added the creation date of the spider based on the last deploy asociated with this spider

* Modify creation date spider

In meeting with my reviewer we agree on creation date of a spider, it must be oldest deploy associated with that spider. Additional information of my changes are: modified code for getting oldest creation date of a deploy associated to my spider, delete REACT_APP_API_BASE_URL variable on .env.development file and delete property output in webpack configuration file.

* Add List of Cards and List of Rows components

I added the above section of the mockup, the first two Cards and the List of Rows that show general information about the spider, the calls to the API are in progress

* Add SpiderJobs Data and filtering button

Added SpiderJobs additional information such as tags and arguments, checking estela api and status of Jobs to organize different tables for jobs

* Ignore some content in the mockup

In meetings with my reviewer Raymond, we agreed on ignoring some fields that are not required yet in the spider detail mockup (such as related scheduled jobs) or should be implemented in the backend side (such as creation date field in the Spider model)

* Issue on server MIME type error

* Add Tables logic in overview page

I added the tables logic in the overview page to see queued, running and completed jobs, also in a meeting with my reviewer Raymond we noticed an issue between estelaCLI and estela web interface we will fix it soon.

* Add Spider creation date based on deploy

Added the creation date of the spider based on the last deploy asociated with this spider

* Modify creation date spider

In meeting with my reviewer we agree on creation date of a spider, it must be oldest deploy associated with that spider. Additional information of my changes are: modified code for getting oldest creation date of a deploy associated to my spider, delete REACT_APP_API_BASE_URL variable on .env.development file and delete property output in webpack configuration file.

* Add changes from pull request

Adding the changes from the comments of my reviewer in the pull request, changes made are: getting scheduled jobs number from the jobs cronjobs associated, changed custom colors from defined colors on tailwind.config file, disabled all buttons that do not have any logic, deleted .env.development file.

* Untrack files contained in the -gitignore files

* Remove development logs

Removed some debugging logs ind evelopment (bad practice sorry)